pdf_summarizer_agent/backend/pyproject.toml
[project]
name = "pdf-summarizer-backend"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn",
  "langgraph",
  "langchain-core",
  "google-generativeai",
  "pypdf",
  "requests",
  "beautifulsoup4",
  "python-dotenv"
]
